refactor: Split out ops/steps
