+++ src/ops/mod.rs
@@ -1,0 +1,6 @@
+pub mod cargo;
+pub mod cmd;
+pub mod git;
+pub mod replace;
+pub mod shell;
+pub mod version;
