+++ src/ops/shell.rs
